CustomTabBar almost finished
